requirements.txt
fastapi
uvicorn[standard]
openai
pyyaml
requests
chromadb
sentence-transformers
scikit-learn
